debugging Database complete

+++ app/src/main/java/com/crost/aurorabzalarm/data/model/AceMagnetometerDataModel.kt
@@ -5,12 +5,12 @@ import androidx.room.PrimaryKey
 import com.crost.aurorabzalarm.data.ParserConstants.ACE_TABLE_NAME
 
 @Entity(tableName = ACE_TABLE_NAME)
-data class AceMagnetometerDataModel (
+data class AceMagnetometerDataModel(
     @PrimaryKey(autoGenerate = false)
     var datetime: Long,
-    var bx: Float,
-    var by: Float,
-    var bz: Float,
-    var bt: Float,
+    var bx: Double,
+    var by: Double,
+    var bz: Double,
+    var bt: Double,
 )
